Break the cards<->pbn package cycle; extract DDResult to its own file
Mapped the internal package dependency graph from actual imports and
found two real 2-cycles: cards<->gambit and cards<->pbn, both traced
to Whist.scala's @main def doubleDummySolver, which lived in cards but
needed gambit (to run a search) and pbn (to parse the input file).

Moved the @main to a new com.phasmidsoftware.bridge.DoubleDummySolver,
at the bare package root above cards/gambit/pbn -- this cleanly breaks
cards<->pbn (pbn->cards is now the only edge in that pair).

cards<->gambit remains, and isn't a quick fix: State (which must stay
in cards, since Deal/Holding/Hand all need Trick for core domain logic,
and Trick.enumerateSubsequentPlays itself needs Whist) has whist: Whist
as an actual constructor field, not just an incidental reference.
Relocating Whist alone (tried first, then reverted) just moves the
same cycle down to State/Trick instead of removing it. A real fix
would mean changing State's constructor to hold deal/openingLeader/
strain directly instead of embedding Whist -- a data-model change
rippling through every State/Whist call site in the old engine and its
tests, out of proportion to this cleanup.

Also split DDResult (used broadly by both engines, not Whist-specific)
out of Whist.scala into its own cards/DDResult.scala, matching the
Contract.scala precedent, and dropped a redundant self-import in
Deal.scala noticed along the way.

Full suite (406 tests) passes; IT/compile succeeds; spot-checked
WinchesterBoard1FuncSpec still produces the identical result. No
behavior changes -- organization only.
